requirements.txt
streamlit
pandas
folium
streamlit-folium
fpdf2
matplotlib
sqlalchemy